+++ src/admin/listeners.js
@@ -7,7 +7,7 @@ import {
 
 import { PERSONALITY } from "../personality.js";
 import {
-  checkProdTestMode,
+  isTestServer,
   clientEventUpdateProcess,
   endCasesEmbed,
   fetchAuditLog,
@@ -121,7 +121,7 @@ export const onChannelUpdate = async (oldChannel, newChannel) => {
 
   //basic operations
   const logChannel = await getLogChannel(newChannel); //get logChannelId
-  if (process.env.DEBUG === "no" && checkProdTestMode(logChannel)) return; //if in prod && modif in test server
+  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
   const embed = setupEmbed("DARK_AQUA", chnUp, newChannel, "tag"); //setup embed
   const chnLog = await fetchAuditLog(oldChannel.guild, "CHANNEL_UPDATE", 1); //get auditLog
 
@@ -217,20 +217,15 @@ export const onChannelUpdate = async (oldChannel, newChannel) => {
           : await newChannel.guild.roles.fetch(cur[0].id);
 
       //write text
-      const textAdded = "\n" + added.join("\n");
-      const textRemoved = "\n" + removed.join("\n");
-      return (
-        acc +
-        "\n" +
-        obj.toString() +
-        "\n" +
-        perm.permAdded +
-        textAdded +
-        "\n" +
-        perm.permRemoved +
-        textRemoved +
-        "\n"
-      );
+      const textAdded =
+        added.length !== 0
+          ? "\n" + perm.permAdded + "\n" + added.join("\n")
+          : "";
+      const textRemoved =
+        removed.length !== 0
+          ? "\n" + perm.permRemoved + "\n" + removed.join("\n")
+          : "";
+      return acc + "\n" + obj.toString() + textAdded + textRemoved;
     }, "");
 
     if (modifs.length !== 0) {
@@ -308,10 +303,18 @@ export const onThreadCreate = async (thread, newly) => {
   if (thread) {
     //sometimes thread is null
     if (thread.joinable && !thread.joined) await thread.join(); //join thread created
+    const logChannel = await getLogChannel(thread); //get logChannel
+    if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
 
-    const logType = "THREAD_CREATE";
-    const perso = "threadCreate";
-    generalEmbed(perso, thread, "DARK_GREY", logType, 1, null, "tag");
+    const perso = PERSONALITY.getAdmin().threadCreate;
+    const log = await fetchAuditLog(thread.guild, "THREAD_CREATE", 1); //get auditLog
+    const executor = log.executor
+      ? log.executor
+      : await thread.guild.members.fetch(thread.ownerId);
+    const embed = setupEmbed("DARK_GREY", perso, thread, "tag"); //setup embed
+    console.log("log.executor", log.executor.id);
+
+    finishEmbed(perso, executor, embed, logChannel);
   } else console.log("threadCreateIsNull", thread, newly);
 };
 
@@ -355,7 +358,7 @@ export const onRoleUpdate = async (oldRole, newRole) => {
   const auditLog = personality.auditLog;
 
   const logChannel = await getLogChannel(newRole); //get logChannelId
-  if (process.env.DEBUG === "no" && checkProdTestMode(logChannel)) return; //if in prod && modif in test server
+  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
   const embed = setupEmbed("DARK_GOLD", roleUp, newRole); //setup embed
 
   //get client
@@ -445,7 +448,7 @@ export const onMessageDelete = async (message) => {
   const auditLog = personality.auditLog;
 
   const logChannel = await getLogChannel(message, "thread"); //get logChannel
-  if (process.env.DEBUG === "no" && checkProdTestMode(logChannel)) return; //if in prod && modif in test server
+  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
 
   const uDate = new Date(message.createdAt); //set date as Date object
   if (currentServer.name === "prod") uDate.setHours(uDate.getHours() + 1); //add 1h to date
@@ -586,7 +589,7 @@ export const onMessageUpdate = async (oldMessage, newMessage) => {
   const auditLog = personality.auditLog;
 
   const logChannel = await getLogChannel(nMessage, "thread"); //get logChannel
-  if (process.env.DEBUG === "no" && checkProdTestMode(logChannel)) return; //if in prod && modif in test server
+  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
 
   const embed = setupEmbed("DARK_GREEN", messageU, nMessage.author, "tag"); //setup embed
   //no auditLog when message update
@@ -596,14 +599,18 @@ export const onMessageUpdate = async (oldMessage, newMessage) => {
     const unpinLog = await fetchAuditLog(nMessage.guild, "MESSAGE_UNPIN", 1); //get auditLog
     const unpinned = messageU.unpinned;
     embed.addField(unpinned.title, unpinned.text, true); //add unpinned text
+    embed.addField(messageU.channel, `<#${oMessage.channelId}>`, true); //message channel
 
     //add message link
     const link = `[${messageU.linkMessage}](${nMessage.url})`;
-    embed.addField(messageU.linkName, link);
+    embed.addField(messageU.linkName, link, true);
+
+    //add executor
+    embed.addFields({name: unpinned.executor, value: unpinLog.executor.toString(), inline: true})
 
     const messageList = await endCasesEmbed(
       nMessage,
-      unpinLog,
+      null,
       messageU,
       auditLog,
       embed,
@@ -618,14 +625,18 @@ export const onMessageUpdate = async (oldMessage, newMessage) => {
     const pinLog = await fetchAuditLog(nMessage.guild, "MESSAGE_PIN", 1); //get auditLog
     const pinned = messageU.pinned;
     embed.addField(pinned.title, pinned.text, true); //add unpinned text
+    embed.addField(messageU.channel, `<#${oMessage.channelId}>`, true); //message channel
 
     //add message link
     const link = `[${messageU.linkMessage}](${nMessage.url})`;
     embed.addField(messageU.linkName, link, true);
 
+    //add executor
+    embed.addFields({name: pinned.executor, value: pinLog.executor.toString(), inline: true})
+
     const messageList = await endCasesEmbed(
       nMessage,
-      pinLog,
+      null,
       messageU,
       auditLog,
       embed,
@@ -761,7 +772,7 @@ export const onGuildMemberUpdate = async (oldMember, newMember) => {
   const auditLog = personality.auditLog;
 
   const logChannel = await getLogChannel(newMember); //get logChannel
-  if (process.env.DEBUG === "no" && checkProdTestMode(logChannel)) return; //if in prod && modif in test server
+  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
   const embed = setupEmbed("ORANGE", timeout, user, "tag"); //setup embed
   const timeoutLog = await fetchAuditLog(newMember.guild, "MEMBER_UPDATE", 1); //get auditLog
   const reason = timeoutLog.reason; //get ban reason
@@ -785,7 +796,7 @@ export const onGuildMemberRemove = async (memberKick) => {
   const auditLog = personality.auditLog;
 
   const logChannel = await getLogChannel(memberKick); //get logChannel
-  if (process.env.DEBUG === "no" && checkProdTestMode(logChannel)) return; //if in prod && modif in test server
+  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
   const kickLog = await fetchAuditLog(memberKick.guild, "MEMBER_KICK", 1); //get auditLog
   const reason = kickLog ? kickLog.reason : null; //get kick reason
 
@@ -840,10 +851,14 @@ export const onGuildMemberRemove = async (memberKick) => {
 };
 
 export const onGuildMemberAdd = async (guildMember) => {
-  console.log("onGuildMemberAdd", guildMember.displayName);
+  const currentServer = COMMONS.fetchGuildId(guildMember.guild.id);
 
-  const db = guildMember.client.db;
-  const authorId = guildMember.id;
-  const date = guildMember.joinedAt.toISOString();
-  addAlavirien(db, authorId, 0, date);
+  if (currentServer.name === "prod" && process.env.DEBUG === "no") {
+    console.log("onGuildMemberAdd", guildMember.displayName);
+
+    const db = guildMember.client.db;
+    const authorId = guildMember.id;
+    const date = guildMember.joinedAt.toISOString();
+    addAlavirien(db, authorId, 0, date);
+  }
 };