+++ src/personality.js
@@ -4,15 +4,17 @@ const path = "static/personalities/";
 const admin = JSON.parse(readFileSync(path + "admin.json"));
 const announces = JSON.parse(readFileSync(path + "announces.json"));
 const personalities = JSON.parse(readFileSync(path + "personalities.json"));
+const colors = JSON.parse(readFileSync("static/colors.json"));
 
 //export var PERSONALITY = personalities.normal; // common var for all files
 
 class Personality {
-  constructor(name, personality, admin, announces) {
+  constructor(name, personality, admin, announces, colors) {
     this.name = name;
     this.personality = personality;
     this.admin = admin;
     this.announces = announces;
+    this.colors = colors;
   }
 
   set(name, personality) {
@@ -37,13 +39,17 @@ class Personality {
   getAnnounces() {
     return this.announces;
   }
+  getColors() {
+    return this.colors;
+  }
 }
 
 export const PERSONALITY = new Personality(
   "normal",
   personalities.normal,
   admin,
-  announces
+  announces,
+  colors
 );
 
 // coming soon, but not today